Foram desabilidado argumentos para desabilitar funcionalidades
related #21

+++ helprs/restfy.py
@@ -1,11 +1,13 @@
 import json
 
-from django.http import HttpResponse
+
 from django.db import transaction
 from django.db.models import Q
+from django.http import HttpResponse, HttpResponseNotAllowed
 
 
-def make_rest(Serializer):
+def make_rest(Serializer, allow_list=True, allow_get=True, allow_create=True, 
+        allow_update=True, allow_delete=True):
 
     Model = Serializer.Model()
 
@@ -89,7 +91,7 @@ def make_rest(Serializer):
                 response = HttpResponse(
                     content_type='application/json',
                     content=json.dumps([
-                        Serializer.encode(state) for state in query
+                        Serializer.encode(instance) for instance in query
                     ])
                 )
             else:
@@ -189,22 +191,41 @@ def make_rest(Serializer):
 
         response = None
 
-        if request.method == 'GET':
+        if allow_list and request.method == 'GET':
             response = _list(request)
 
-        elif request.method == 'POST':
+        elif allow_create and request.method == 'POST':
             response = _create(request)
+        else:
+            permitted_methods = []
 
+            allow_list and permitted_methods.append('GET')
+            allow_create and permitted_methods('POST')
+
+            response = HttpResponseNotAllowed(
+                permitted_methods
+            )
         return response
 
     def _by_id(request, id):
-        if request.method == 'GET':
+        if allow_get and request.method == 'GET':
             return _get_by_id(request, id)
 
-        elif request.method == 'DELETE':
+        elif allow_delete and request.method == 'DELETE':
             return _delete_by_id(request, id)
 
-        elif request.method == 'PUT':
+        elif allow_update and request.method == 'PUT':
             return _update_by_id(request, id)
+        else:
+
+            permitted_methods = []
+
+            allow_get and permitted_methods.append('GET')
+            allow_delete and permitted_methods('DELETE')
+            allow_update and permitted_methods('PUT')
+
+            return HttpResponseNotAllowed(
+                permitted_methods
+            )
 
     return _index, _by_id
